ベースクラスの追加
