+++ app/controllers/api/api_controller.rb
@@ -2,20 +2,5 @@
 
 module Api
   class ApiController < ActionController::API
-    include ActionController::HttpAuthentication::Token::ControllerMethods
-    before_action :require_access_token
-
-    protected
-
-    def require_access_token
-      access_token = request.headers['Authorization']
-      return render json: { message: 'Missing authentication token' }, status: :unauthorized if access_token.nil?
-
-      begin
-        AccessToken.from_token(access_token[7..])
-      rescue JWT::DecodeError
-        render json: { message: 'Unauthorized. Invalid token' }, status: :unauthorized
-      end
-    end
   end
 end
